refactor(supabase): extract duplicated auth, image upload, revalidate and proxy logic
- Move getCurrentUser / getOwnerProfile (with "use cache: private") to
  @ykzts/supabase/auth for sharing between admin and memo.
- Move uploadImage, getImageDimensions and shared image validation
  (MAX_FILE_SIZE, ALLOWED_TYPES, validateImageFile) to
  @ykzts/supabase/image-upload. Update consumers in forms and upload
  actions (avatar, key-visual).
- Extract revalidate handler (schema + secret validation + response)
  to @ykzts/supabase/revalidate. Thin out the identical route handlers
  in blog and portfolio.
- Move updateSession proxy logic to @ykzts/supabase/proxy (renamed
  from ./middleware for consistency with Next.js 16 "proxy" naming
  and the repo's existing proxy.ts convention in admin/blog/memo).
- Extract extractTextFromPortableText to @ykzts/utils/portable-text
  (used by admin embeddings generation).
- Deduplicate date URL formatting in admin by reusing
  @ykzts/utils/blog-urls.
- Remove local duplicate files (auth.ts, upload-image.ts,
  lib/supabase/proxy.ts) from admin and memo.
- Update all import sites and related code (data, posts, uploads,
  layouts, actions, components, tests).
- Minor naming alignments (handleRevalidatePOST -> handleRevalidate).

Addresses duplication identified in #4016 and subsequent review of
similar patterns across the monorepo (Supabase clients, owner checks,
validation, query helpers, etc.).

Closes #4016
