+++ apps/admin/app/(authenticated)/works/_components/work-form.tsx
@@ -15,6 +15,7 @@ import {
   verticalListSortingStrategy,
 } from "@dnd-kit/sortable";
 import { RichTextEditor } from "@ykzts/editor";
+import { uploadImage } from "@ykzts/supabase/image-upload";
 import { Alert, AlertDescription } from "@ykzts/ui/components/alert";
 import { Button } from "@ykzts/ui/components/button";
 import {
@@ -32,7 +33,6 @@ import {
 import type { MouseEvent } from "react";
 import { useActionState, useCallback, useState } from "react";
 import { generateUniqueSlugForWork } from "@/lib/slug";
-import { uploadImage } from "@/lib/upload-image";
 import { generateSlug } from "@/lib/utils";
 import { SortableItem } from "./sortable-item";
