+++ apps/admin/proxy.ts
@@ -1,5 +1,5 @@
+import { updateSession } from "@ykzts/supabase/proxy";
 import type { NextRequest } from "next/server";
-import { updateSession } from "@/lib/supabase/proxy";
 
 export async function proxy(request: NextRequest) {
   return await updateSession(request);
