+++ apps/memo/app/[...path]/actions.ts
@@ -1,9 +1,9 @@
 "use server";
 
+import { getOwnerProfile } from "@ykzts/supabase/auth";
 import { createServerClient } from "@ykzts/supabase/server";
 import type { Json } from "@ykzts/supabase/types";
 import { revalidatePath } from "next/cache";
-import { getOwnerProfile } from "@/lib/auth";
 
 export type UpdateMemoState = {
   error?: string;
